requirements.txt
python-binance
requests
Flask 